refactor(data-quality): align frontend module boundaries
