--- yak-ops-ui/src/pages/data-quality/execution/service.ts
@@ -1,2 +1,0 @@
-/** @deprecated Import direct data functions from `@/services/data-quality`. */
-export { qualityExecutionWorkspaceApi } from '@/services/data-quality/execution-legacy';
