+++ yak-ops-ui/src/pages/data-quality/table-config/components/DataSourceTreePane.tsx
@@ -14,12 +14,8 @@ import { useMemo, type PointerEvent as ReactPointerEvent } from 'react';
 import type { QualityDataSourceNode } from '../types';
 import { groupQualityDataSourceNodes } from '../utils';
 
-const EMPTY_SOURCE_NODES: QualityDataSourceNode[] = [];
-
 interface DataSourceTreePaneProps {
-  sourceNodes?: QualityDataSourceNode[];
-  /** Compatibility with callers that still use the pre-refactor prop name. */
-  treeData?: QualityDataSourceNode[];
+  sourceNodes: QualityDataSourceNode[];
   treeLoading: boolean;
   selectedNodeKey?: string;
   leftWidth: number;
@@ -39,7 +35,6 @@ interface QualityTreeDataNode extends DataNode {
 
 const DataSourceTreePane = ({
   sourceNodes,
-  treeData: legacyTreeData,
   treeLoading,
   selectedNodeKey,
   leftWidth,
@@ -48,12 +43,9 @@ const DataSourceTreePane = ({
   onResizeStart,
   onCollapsedChange,
 }: DataSourceTreePaneProps) => {
-  const resolvedSourceNodes =
-    sourceNodes ?? legacyTreeData ?? EMPTY_SOURCE_NODES;
-
   const treeData = useMemo<QualityTreeDataNode[]>(
     () =>
-      groupQualityDataSourceNodes(resolvedSourceNodes).map((group) => ({
+      groupQualityDataSourceNodes(sourceNodes).map((group) => ({
         key: `type:${group.dataSourceType}`,
         title: group.dataSourceType,
         kind: 'group',
@@ -68,7 +60,7 @@ const DataSourceTreePane = ({
           isLeaf: true,
         })),
       })),
-    [resolvedSourceNodes],
+    [sourceNodes],
   );
 
   const renderTitle: TreeProps['titleRender'] = (rawNode) => {
