+++ yak-ops-ui/src/services/data-source/index.ts
@@ -1,3 +1,4 @@
 export * from './api';
+export * from './catalog';
 export * from './driver';
 export type * from './types';
